requirements.txt
asyncio
numpy
pyaudio
torch
torchaudio
openai
python-dotenv
colorama